Exercise submission by bejumi
My goal was to maintain some of the default styling of the site but move
all the functionality to the map itself.
The only additional external library I included was Bootstrap to make
the page look a bit more consistent as well as free airport pin icon
from flaticonthat is credited on the page. See the README.md for
complete details.