If a user selects an airport that is not currently shown on the map, how do they find it? Can you make this easier?
Using some useful googled code, when the user selects an airport, we will recenter on it (at the current zoom level). Also added feature so that if the users clicks "show all markers", we will reset the map's size so that we can see all of the markers; kinda cool!  Thanks, Internet!