Updated implementation to meet 5/6 requirements.
Updated the code to fill in the remainder of the data for a location (1), allow a user to remove a location by selecting the marker on the map (2), and pre-sort the list to make navigating the list easier (3).

Added code to lazily initialize markers and store them in an object (hash) so that they are created at most one time (5), and select an re-center the map if it is not within the bounds of the current map (6).

Regarding #4, one method of improving the user experience might be to create a series of UI cards that contain all the information, which can be accessed by selecting a value from a "selected items" list. The workflow is something like this:

1. The user selects an airport in the options list
2. The selection is added to a _new_ options list, which stores only the sites a user has selected
   2a. If the selection already exists in the set, then it is not added
3. To display the card(s), the user selects one (or more) options from the visited (selected) items list, and then selects an option to show more data. On the resultant screen, the user can see the full information about the selection they've made, instead of just the resource key (airport code) that was used for the selection.

Because this method results in having a lot of data the user may not care about, a button can be included to add the item to a selection list, which allows the user to retain only the items they're interested in (similar to an Amazon shopping cart).